requirements.txt
apispec==5.2.2
Brotli==1.0.9
cachelib==0.8.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
Flask==2.1.2
flask-apispec==0.11.1
Flask-Caching==1.11.1
Flask-Compress==1.12
Flask-Cors==3.0.10
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.16.0
numpy==1.21.6
packaging==21.3
pymongo==4.1.1
pyparsing==3.0.9
python-json-logger==2.0.2
PyYAML==6.0
requests==2.28.0
simplejson==3.17.6
six==1.16.0
typing_extensions==4.2.0
urllib3==1.26.9
webargs==8.1.0
Werkzeug==2.1.2
zipp==3.8.0
